migrate python 2.7 to python 3.5.2

+++ classes/RGBImageSlicer.py
@@ -16,10 +16,10 @@ class RGBImageSlicer:
     @property
     def first_quarter(self):
         width_from = 0
-        width_to = self.width / 2
+        width_to = int(self.width / 2)
 
         height_from = 0
-        height_to = self.height / 2
+        height_to = int(self.height / 2)
 
         first_quarter_image = self.image[height_from:height_to, width_from:width_to, :]
 
@@ -27,11 +27,11 @@ class RGBImageSlicer:
 
     @property
     def second_quarter(self):
-        width_from = self.width / 2
+        width_from = int(self.width / 2)
         width_to = self.width
 
         height_from = 0
-        height_to = self.height / 2
+        height_to = int(self.height / 2)
 
         second_quarter_image = self.image[height_from:height_to, width_from:width_to, :]
 
@@ -40,9 +40,9 @@ class RGBImageSlicer:
     @property
     def third_quarter(self):
         width_from = 0
-        width_to = self.width / 2
+        width_to = int(self.width / 2)
 
-        height_from = self.height / 2
+        height_from = int(self.height / 2)
         height_to = self.height
 
         third_quarter_image = self.image[height_from:height_to, width_from:width_to, :]
@@ -51,10 +51,10 @@ class RGBImageSlicer:
 
     @property
     def fourth_quarter(self):
-        width_from = self.width / 2
+        width_from = int(self.width / 2)
         width_to = self.width
 
-        height_from = self.height / 2
+        height_from = int(self.height / 2)
         height_to = self.height
 
         fourth_quarter_image = self.image[height_from:height_to, width_from:width_to, :]
